[Refactor] HubRoute의 이동 거리 타입 변경 (double → int)

+++ src/main/java/org/gupang/hub/application/route/dto/HubRouteInfo.java
@@ -9,7 +9,7 @@ public record HubRouteInfo(
         String startHubName,
         UUID endHubId,
         String endHubName,
-        double estimatedDistance,
+        int estimatedDistance,
         int estimatedDuration
 ) {
